[add] 카드 수정 로직 추가

+++ app/src/main/java/com/toyou/toyouandroid/data/home/dto/response/CardDetail.kt
@@ -11,7 +11,7 @@ data class CardDetail(
 )
 
 data class AnswerQuestion(
-    @SerializedName("questionId") val id : Int,
+    @SerializedName("questionId") val id : Long,
     @SerializedName("content") val content : String,
     @SerializedName("questionType") val type: String,
     @SerializedName("questioner") val questioner : String,
